+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.1a4 (24.04.2020)
+### Bugfixes
+* fixed auto-generation of `DATE` based on the file date
+### Feature
+* changed `AlternateAlignment` to `SectionAlignment` with values `alternate`, `center`, `left` and `right`
+* `SectionAlignment` can now be overwritten by individual section content files as well
+* added support for `SectionBackground` with values `alternate`, `dark` and `primary`
+* added support for `ReplaceHeading`
+* menu items linking to the current page with a fragment now use the `js-scroll-trigger` CSS class
+
 ## 0.1a3 (21.04.2020)
 ### Feature
 * added support for pages with section